requirements.txt
python-dotenv
pandas
numpy
setuptools
certifi
pymongo[srv]
